apple 영수증 데이터 long 타입 변경

+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/payment/service/utils/dto/response/Receipt.java
@@ -18,7 +18,7 @@ public class Receipt {
     @JsonAlias("bundle_id")
     private String bundleId;
     @JsonAlias("download_id")
-    private Integer downloadId;
+    private Long downloadId;
     @JsonAlias("in_app")
     private List<InApp> inApp;
     @JsonAlias("original_application_version")
